add comments for value gather

+++ openrlhf/models/reward_model.py
@@ -62,8 +62,17 @@ class RewardModel(nn.Module):
         if self.training:
             reward = values[:, -1]
         else:
-            # assume that there is some padding on both sides
-            eos_indices = attention_mask.size(1) - 1 - attention_mask.fliplr().argmax(dim=1, keepdim=True)
+            # The following code is equivalent to:
+            #
+            # last_value = []
+            # for i in range(sequences.size(0)):
+            #     for t in reversed(range(sequences.size(1))):
+            #         if attention_mask[i][t] > 0.5:
+            #             last_value.append(values[i][t])
+            #             break
+            # reward = torch.stack(last_value)
+            #
+            eos_indices = attention_mask.size(1) - 1 - attention_mask.long().fliplr().argmax(dim=1, keepdim=True)
             reward = values.gather(dim=1, index=eos_indices).squeeze(1)
 
             # normalize reward in eval mode
